feat: Added FPSSxript.cs
Awake Method: This method is called when the script instance is being loaded. It checks if the fpsText component is assigned. If not, it tries to find a TextMeshProUGUI component on the same GameObject. If it can't find one, it logs an error message.

Update Method: This method is called once per frame. It calculates the FPS and updates the fpsText with the current FPS value.

CalculateFPS Method: This method calculates the FPS using a smoothing algorithm. It updates the _deltaTime value with a smoothing factor to make the FPS readings less fluctuating.

DisplayFPS Method: This method displays the calculated FPS value on the assigned TextMeshProUGUI component. It rounds the FPS value and updates the text accordingly.

Overall, the script is used to monitor the game's performance and provide the FPS information to the user.